Lookup of Chinook FRAM (marine) fisheries

Contains FisheryIDs and names from a current Chinook project database and appends several fields of metadata and catch database identifiers. Note that additional TAMM fisheries are required for a complete depiction of impacts (i.e., including terminal and freshwater mortality). Various fields are NA for fisheries for which they do not apply.

lu_chin_fishery

Format

data frame (tbl_df) with 73 rows and 11 cols:

Species

character Chinook

VersionNumber

integer 1

FisheryID

integer 1:73

FisheryName

character short name

FisheryTitle

character long name

catch_soure

character designating primary data source

region

character string of geographic domain

areacode

character string of CRC identifiers

CatchAreaCode

character string of WDFW FishTix identifiers

FisherTypeDescription

character string FishTix Tr/NT levels

gear

character string of net or line

Source

see vignette 'chinook_lookup_creation'

Examples

lu_chin_fishery |>
dplyr::filter(catch_source == "FishTicket/TOCAS") |>
dplyr::select(catch_source, FisheryID, FisheryName, CatchAreaCode, FisherTypeDescription, gear) |>
tidyr::separate_rows(CatchAreaCode)

#> Error in dplyr::filter(lu_chin_fishery, catch_source == "FishTicket/TOCAS"): object 'lu_chin_fishery' not found