Lookup of Chinook FRAM (marine) fisheries

Contains FisheryIDs and names from a current Chinook project database and appends several fields of metadata and catch database identifiers. Note that additional TAMM fisheries are required for a complete depiction of impacts (i.e., including terminal and freshwater mortality). Various fields are NA for fisheries for which they do not apply.

lu_chin_fishery

Format

data frame (tbl_df) with 73 rows and 11 cols:

Species

character Chinook

VersionNumber

integer 1

FisheryID

integer 1:73

FisheryName

character short name

FisheryTitle

character long name

catch_soure

character designating primary data source

region

character string of geographic domain

areacode

character string of CRC identifiers

CatchAreaCode

character string of WDFW FishTix identifiers

FisherTypeDescription

character string FishTix Tr/NT levels

gear

character string of net or line

Source

see vignette 'chinook_lookup_creation'

Examples

lu_chin_fishery |>
dplyr::filter(catch_source == "FishTicket/TOCAS") |>
dplyr::select(catch_source, FisheryID, FisheryName, CatchAreaCode, FisherTypeDescription, gear) |>
tidyr::separate_rows(CatchAreaCode)
#> # A tibble: 84 x 6
#>    catch_source     FisheryID FisheryName CatchAreaCode FisherTypeDescrip~ gear 
#>    <chr>                <int> <chr>       <chr>         <chr>              <chr>
#>  1 FishTicket/TOCAS        23 NT GHb Net  02A           Non-Treaty         net  
#>  2 FishTicket/TOCAS        23 NT GHb Net  02B           Non-Treaty         net  
#>  3 FishTicket/TOCAS        23 NT GHb Net  02C           Non-Treaty         net  
#>  4 FishTicket/TOCAS        23 NT GHb Net  02D           Non-Treaty         net  
#>  5 FishTicket/TOCAS        24 Tr GHb Net  02A           Treaty             net  
#>  6 FishTicket/TOCAS        24 Tr GHb Net  02B           Treaty             net  
#>  7 FishTicket/TOCAS        24 Tr GHb Net  02C           Treaty             net  
#>  8 FishTicket/TOCAS        24 Tr GHb Net  02D           Treaty             net  
#>  9 FishTicket/TOCAS        25 WillapaNet  02G           Non-Treaty         net  
#> 10 FishTicket/TOCAS        25 WillapaNet  02H           Non-Treaty         net  
#> # ... with 74 more rows